Create an encapsulated Series object
As data enters the DataProvider, convert it into an instance of a Series
object. This will contain all of the metadata as well as the functions
that callers will use down the line, like scale calculations.

This also allows the callers to have the metadata live directly with the
series object passed in from the loader, instead of as a series of
disconnected metadata objects (eg: hiddenSeries, colors, strokeWidths).
Backwards compatibility is maintained by mixing the metadata arrays in
the DataProvider, with the series object's metadata taking precedence.

Finally, create an AxisCollection object and extract a dedicaetd YAxis
object. These maintain single-responsibility behavior, where they are
only responsible for rendering their level of data. This gives us much
more manageable code in LineChart and Axis. It also allows for an
<AxisCollection> to be added as a sibling of <LineChart>, but this has
not yet been implemented because it causes issues with calculating the
chart width once the axes become dynamically shown and hidden.