Clean up bugs found in the 0.2.0 migration
Calculate domain based on y0/y1

If y0Accessor and y1Accessor are specified, use those to calculate the
domain for the chart.

Remove styled-components

This has disastrous consequences if the user of griff also uses
styled-components, and with a different version (virtually guaranteed).

This change also involved an attempt to move to CSS Grid to
automatically determine the components sizes, but without relying on
sizeMe. However, this was aborted for the time being, with some small
sizing remnants hiding in the code. One example is the new contextChart
config, which looks like this:

  contextChart: PropTypes.shape({
    // Whether to show the context chart at all.
    visible: PropTypes.bool,
    // How high (in pixels) the context chart should be. Note that an
    // extra 50 pixels will be added for its X axis.
    height: PropTypes.number,
    // How high (in percent of overall height) the context chart should
    // be. Note that 50 pixels is reserved for it X axis height.
    heightPct: PropTypes.number,
  }