Add AxisDisplayMode functionality (#56)
Add support for several y-axis display modes:

  ALL: Show all of the axes.
  NONE: Don't show any of the axes.
  COLLAPSED: Hide the real axes and instead render a single generic axis
             without values or color.

The yAxisDisplayMode property can be specified on an individual series
object, too. In this case, it overrides the one set on the LineChart
component, by rules of specificity.

This also adds event listeners to AxisCollection, as specified via
onAxisMouseEnter / onAxisMouseExit properties on LineChart.