Got a combined y axis working
There's one bit of weirdness though: if you zoom the y axis and then
zoom the chart, the zoom is *not* respected. However, if you do it the
other way around (zoom the chart first, then the axis) then the zoom is
respected.

I think it's because we're only setting one transformation, but I'm not
sure. I will continue to investigate.