Correctly render the number of points
The randomData generator shouldn't guess at how many points it should
use; it should generate exactly as many as requested.

Also handle if scatterplot points have the same timestamp -- if they do,
we don't need to double-up on the number of points rendered. There's
even room here for allowing a threshold, but we can leave that logic up
to the specific loader.