Document the branch layout & versioning system (#116) (#118)
Document how the branches play together, what the versions will be, and
how to make changes to griff. This also adds a script (run with `yarn
release`) to determine the new version number, publish a release, and
push the tag to GitHub.

This is a cherry-pick of 4cfe5bf and sets the version to `0.3.0-0` for
pre-release builds.