Add support for series collections
Series may now be grouped into collections, which are passed into the
DataProvider by the new `collections` property. Series which declare
that they are a member of a given collection (by their `collectionId`
property) are not rendered as an individual axis, and are instead folded
into a unified axis with combined zoom & translation capabilities.