Properly scale the collections
It turns out that there are no shortcuts allowed -- the collection needs
to be properly scaled. We can't just lazily compute the yDomain from its
component series.

When items (series, collections) leave DataProvider, they will all have
yDomain properties and be fully ready to scale.

The yDomains are not applied to the context chart, unless they're
specified from the get-go. I believe that this is a logical way to do
things, but I'm open to other views.