Support configurable axis placements
Add a new `yAxisPlacement` property to both LineChart and series. This
value is one of the defined `AxisPlacement` constants, and it controls
where the respective axes/axis live/s.

It has the following behaviors:
  UNSPECIFIED
    Same as not using the property, and the default value (RIGHT) is
    used.
  RIGHT
    Default behavior -- the y-axes are rendered to the right of the
    line chart.
  LEFT
    Axes are rendered to the left of the line chart; series are mirrored
    so that the proximity to the chart is preserved.
  BOTH
    Axes are rendered on both sides of the chart. These are kept in sync
    with each other.

If the series objects specify this option, then they will be placed as
specified, overriding the chart's value if necessary.

Fixes #89