Add support for grid lines
Expose a `grid` property on the LineChart which controls the grid line
specification. This object is documented in `proptypes.js`, but here's a
summary:

The `y` property of the grid specification controls how horizontal
lines behave (they're distributed along the Y values). These can either
be rendered statically or linked to one or more series. If there are
links, then the grid lines will be transformed just like the y-axis tick
marks as the data is scaled.

The `x` property of the grid specification controls how vertical lines
behave (they're distributed along the X values). Similarly, these can
either be rendered statically or linked to the time value. There is no
series linkage necessary for this, because the x-axis is independent of
the series being plotted.

Fixes #113