Don't use a class for static constant holders (#135)
Instead of using a class with only static properties, export an object
from `proptypes.js`. No only is this more efficient, but it would
otherwise crash. For example, `GriffPropTypes.collections` was referring
to `GriffPropTypes.collection`, but `GriffPropType` was in the process
of being defined. By moving to a collection of consts, this problem is
avoided.