Add a notion of ySubDomain (#139)
Support the y axis having a subdomain property, much like the x axis
does. In the interest of maintaining backwards compatibility, the
property was not renamed, but the equivalence is:

  baseDomain => yDomain
  subDomain  => ySubDomain

This can be specified on a series, collection, or on the <DataProvider>
element. The ySubDomain property is not respected on the context chart,
but otherwise replaces the yDomain property everywhere else.