Allow collections to be collapsed
Correctly handle collections being collapsed, as well as having their
component parts specify that they'd like to be collapsed, without
falling apart.

This is done by ignoring the yAxisDisplayMode on the component series if
it is placed in a collection.

This also contains a nontrivial performance boost with respect to
collections by using more efficient array methods to require fewer
iterations through the collections.