Introduce scatterplots (#120)
Introduce scatterplots to griff through the introduction of a new
<Scatterplot> component type. This is very similar to <LineChart>, with
some minor behavioral differences:

 - Axes do not individually zoom.
 - x-axis does not have a notion of time, so its values are rendered as
   numbers instead.
 - Series can provide accessors for pointWidth and opacity.
 - Subdomain does not get adjusted on zoom (because time is not a axis
   value).

Data is provided by the same loader object as line charts. In order to
plot two series against each other, an interpolating composite loader
needs to be provided by the calling application.

Lots of example stories can be found in the storybook.