Add point-hovering capabilities
Add a callback function to the series object: onPointHover. This is
called whenever the point is hovered (or un-hovered) to allow the
application to respond accordingly.

This change also puts the points on top of the InteractionLayer for
scatterplots. Without this change, the hovering functionality is blocked
by the InteractionLayer. However, zooming is blocked when on top of a
point.