Copy the series properies after the loader props
When the loader returns data, it gets copied into the enriched series.
This needs to happen before copying the series information from the
props, or else old information will be retained and clobber the
potentiall-changed data coming in from the props.

This was never caught because the storybook loader does not spread the
`oldSeries` object, and instead only returns the `data` property.