Introduce xDomain and xSubDomain
Instead of using a notion of baseDomain and subDomain, introduce a
xDomain and xSubDomain. These mirror the yDomain and ySubDomain
properties. This change is necessary because, with the addition of
scatterplots, not all charted data has a relevant time property.

This will also let the X values be populated by the extent of the data,
similar to the yDomain is calculated by the range of the data.

The xDomain is still passed into the loader so that it can do the right
thing -- for time-based loaders, the xDomain will continue to represent
time. However, for loaders which have time as an independent axis, they
can manage their own notion of which data to load.