Extract XAxis and ContextChart
Make the ContextChart and XAxis components top-level components. This
enables them to be exist independently, as long as they are children of
DataProvider.

For upgrade friction, the components still maintain their own
ContextChart and XAxis components. However, this will be a breaking
change in the 0.4.0 version of this library. Once that happens, callers
will be responsible for adding (and positioning!) their own desired
components. For example:

```
  <DataProvider ... />
    <div style={{ height: 400, width: 400 }}>
      <LineChart ... />
    </div>
    <XAxis ... />
    <div style={{ height: 150 }}>
      <ContextChart ... />
    </div>
  </DataProvider>
```