Change the Scaler to track {sub,}domains (#191)
Instead of the Scaler tracking the d3 transforms to handle zooming and
panning, track the domains and subdomains for each item. This allows
sub-components to query the domain & subdomain for any given item
(series or collection), and render themselves accordingly.

d3 is still used for the event handling, but the zoom / pan logic has
been moved to be handled entirely inside of the ZoomRect component.

This change will enable multiple components to share responsibility for
adjusting zoom. One example of this could be a scatterplot which allows
zooming on the chart itself, but also allowing the axes to further
manipulate the scaling in either direction.

This change also ends the notion that the x-axis represents time, since
this has not been true since the introduction of scatterplots. Instead,
each series now has three possible axes, each with their own domains and
subdomains: x, y, and time.