Handle missing series in the subdomains (#202)
If the series are updated, it's possible to crash because the subdomain
and domain maps in Scaler have not yet been updated. Example logical
flow:

  1. Scaler initialized with series 1.
  2. Scaler's state updated with:
     domainsByItemId: { [1]: { .. } },
     subDomainsByItemId: { [1]: { .. } },
  3. Series 2 added to DataProvider
  4. Scaler notices the new series and updates the state:
     domainsByItemId: { [1]: { .. }, [2]: { .. } },
     subDomainsByItemId: { [1]: { .. }, [2]: { .. } },

When this happens, no other scaling happens in componentDidUpdate
because everything was reinitialized when there are new series added.

As an additional safety measure, Axes will not crash if an invalid
series is given. Instead, a domain (or subdomain) of `[0, 0]` will be
returned.