Formalize the notion of a placeholder domain
Introduce a placeholder domain (owned by Scaler) which will allow for
children to assume a domain exists and is valid, but the Scaler can know
whether or not there is a valid domain or not.

This allows Scaler to use the computed domains and subdomains when
necessary, and use the specified ones when not.