Rework domains & subdomains for x & y axes (#218)
Make x & y domains behave like time: if there is a domain specified,
then the subdomain cannot go beyond those bounds. If neither are
specified, then the subdomain will default to the extent of the data
(with an open-ended domain).

This also removes the notion of a contextSeries. Instead, have the
context chart (if present) simply transform the series to look how it
wants them to look (static domains, drawPoints=false, etc).

This change also introduces an unconsistency: LineCollection takes in a
yScalerFactory with a different signature than xScalerFactory. This will
be cleaned up in a fast-follow PR (to move xScalerFactory to the same
signature as the new yScalerFactory). This refactor was omitted from
this PR to keep it manageable.