Correctly determine axes for AxisCollection
When there are collections which can be on either side of the chart,
AxisCollection needs to handle this correctly. Previously, the following
situation would arise:

Series:
  { id: 1, collectionId: 'left' }
  { id: 2, collectionId: 'left' }
  { id: 3, collectionId: 'right' }
  { id: 4, collectionId: 'right' }

Collections:
  { id: 'left' }
  { id: 'right' }

With this data set, the left axis would have the `left` collection, but
also series `3` and `4`. Similarly, the right axis would have `right`,
`1`, and `2`.