Export GridLines as a top-level component (#249)
Instead of controlling GridLines through props, allow it to be added
directly to the hierarchy as a child of the desired chart (eg, LineChart
or Scatterplot). This allows for a flatter prop structure as well as
beginning the migration of business logic from inside of Griff back into
the client application.