Export ScalerContext as a top-level API item (#252)
In order to allow calling applications to use the ScalerContext, export
it as a top-level item. As part of this change, perform an audit of
everything that is in the ScalerContext and remove unused items.

With ScalerContext being exposed, Griff can play nicely with third-party
charting solutions! This is demonstrated with an example integration
with plotly.js