fix: falsy 0 value xval/yval bug
Also fixed the PropTypes since xpos/ypos are generated and xval/yval are
expected as properties of the areas.

Since end is not required I fixed the end xpos/ypos not being set when
end point was not given in the area. Set the end to be the maximum
yvalue and maximum x value. Added a story to test for this.

Fixes [OI-1910]

[OI-1910]: https://cognitedata.atlassian.net/browse/OI-1910