Don't require start and set default xpos/ypos
The xpos and ypos are now set individually. If start is undefined set
start as bottom left and if end is undefined set end as top-right.
If start or end is defined but the xpos or ypos is missing use the
default xpos or ypos for start or end appropriately