[OI-1908][OI-1910] Collection support for areas (#334)
* Link areas to collection the series belongs to

Added stories for areas and areas on collections.

Fixes [OI-1908]

[OI-1908]: https://cognitedata.atlassian.net/browse/OI-1908

* fix: falsy 0 value xval/yval bug

Also fixed the PropTypes since xpos/ypos are generated and xval/yval are
expected as properties of the areas.

Since end is not required I fixed the end xpos/ypos not being set when
end point was not given in the area. Set the end to be the maximum
yvalue and maximum x value. Added a story to test for this.

Fixes [OI-1910]

[OI-1910]: https://cognitedata.atlassian.net/browse/OI-1910

* Don't require start and set default xpos/ypos

The xpos and ypos are now set individually. If start is undefined set
start as bottom left and if end is undefined set end as top-right.
If start or end is defined but the xpos or ypos is missing use the
default xpos or ypos for start or end appropriately

* Make area store bounding edges instead of corners

This updates the rendering of Areas and the calculation of area clicks.
It does not change the area prop passed to onAreaDefined.
Updates and fixes stories.